Cooperative IDM

Cooperative IDM

AutonomousMerging.CooperativeIDM — Type.

CooperativeIDM <: DriverModel{LaneFollowingAccel}

The cooperative IDM (c-IDM) is a rule based driver model for merging scenarios. It controls the longitudinal actions of vehicles on the main lane. A cooperation level c controls how the vehicles reacts to the merging vehicle. When c=0 the vehicle ignores the merging vehicle. When c=1 the vehicle considers the merging vehicle as its front car when TTCego > TTCmergin_vehicle. When it is not considering the merging vehicle, the car follows the IntelligentDriverModel.

Fields

- `env::MergingEnvironment = MergingEnvironment(main_lane_angle = 0.0, merge_lane_angle = pi/6)` the merging environment
- `idm::IntelligentDriverModel = IntelligentDriverModel(v_des = env.main_lane_vmax, d_cmf = 2.0, d_max=2.0, T = 1.5, s_min = 2.0, a_max = 2.0)` the default IDM
- `c::Float64 = 0.0` the cooperation level
- `fov::Float64 = 20.0` [m] A field of view, the merging vehicle is not considered if it is further than `fov`

source