Merging MDP

Merging MDP

AutonomousMerging.MergingMDP — Type.

MergingMDP <: MDP{MergingState, Int64}

A discrete state and action MDP representing a merging scenario with only two agents: the ego vehicle on the merge lane and another traffic participant on the main lane.

source